fix: handleConnection 내부에서 disconnect 할 경우 예외처리
내부에서 socket.data에 user가 할당되기 전 disconnect를 할 경우 handleDisconnect에서 다른 행동 없이 종료하도록 예외처리를 했다.

+++ api-server/src/rooms/rooms.gateway.ts
@@ -85,9 +85,12 @@ export class RoomsGateway {
   }
 
   async handleDisconnect(socket: Socket) {
+    if (!socket.data.user) {
+      return;
+    }
+
     this.roomsService.exitRoom(socket, socket.data.roomId);
     this.roomsService.deleteUserSocket(socket.data.user.name);
-
     if (socket.data.roomId) {
       if (socket.data.roomId === 'lobby') {
         this.server.in(socket.data.roomId).emit('user_exit_lobby', {
